refactor: Move login page setup function to helper file
